requirements.txt
streamlit
Pillow
tinify
optimize-images